Day7 (#27)
* Part 1

* Part 2

+++ src/y2019/day5/MachIO.java
@@ -3,22 +3,22 @@ package y2019.day5;
 import y2019.intcode.Machine;
 
 public class MachIO implements Machine.IO {
-    private int nextInput;
-    private int lastOutput;
+    private long nextInput;
+    private long lastOutput;
 
-    public MachIO(int value) {
+    public MachIO(long value) {
         this.nextInput = value;
     }
 
-    public int input() {
+    public long input() {
         return nextInput;
     }
 
-    public void output(int value) {
+    public void output(long value) {
         lastOutput = value;
     }
 
-    public int getLastOutput() {
+    public long getLastOutput() {
         return lastOutput;
     }
 }
