+++ src/y2019/day5/Part1.java
@@ -4,16 +4,16 @@ import utils.Problem;
 import y2019.intcode.Machine;
 import y2019.intcode.Parser;
 
-public class Part1 extends Problem<Integer> {
+public class Part1 extends Problem<Long> {
     private Parser parser;
 
-    public Part1(String fileName, int exp) {
+    public Part1(String fileName, long exp) {
         super(exp);
 
         parser = new Parser(fileName);
     }
 
-    public Integer run() {
+    public Long run() {
         var prog = parser.parse();
         var io = new MachIO(1);
         var mach = new Machine(prog, io);
